Fix act and warning errors on terminal page
