+++ site/src/pages/TerminalPage/types.ts
@@ -1,0 +1,1 @@
+export type TerminalState = "connected" | "disconnected" | "initializing";
